feat: Integrate secure image generation into build process
This change allows for the generation of secure boot-enabled images
using customer-specific keys and secrets. In order to trigger a
build of a secure-boot enabled image, the "KEYS_DIR" environment
variable must be set to a valid, accessible directory which contains
the following files:

- secure-sam-ba-cipher-3.5 (directory containing Secure SAM-BA Cipher tool)
- dev.crt (kernel signing cert)
- dev.key (kernel signing key)
- encrypted_filesystem_key.txt (key used to encrypt rodata)
- license_passcode.txt (sam-ba cipher tool passcode)
- license_private_Prod.pem (sam-ba cipher tool cert)
- license_sama5d3_Prod.txt (sam-ba cipher tool cert for use with SAMA5D3x)
- secure_boot_encryption_key.txt (key used for secure boot)
- secure_mode_sama5d3x.cip (sam-ba cipher tool encrypted command file)
- secure_mode_sama5d3x_nk.cip (sam-ba cipher tool encrypted command file)
- update_signing.crt (SWU update signing cert)
- update_signing.key (SWU update signing key)

Secure boot-enabled images are denoted by the "secureboot" suffix in the
build artifact release tarball filename.

Refs: PROD-16438